+++ dask_ml/cluster/spectral.py
@@ -12,6 +12,7 @@ from scipy.linalg import pinv, svd
 from sklearn.base import BaseEstimator, ClusterMixin
 from sklearn.utils import check_random_state
 
+from .._utils import draw_seed
 from ..metrics.pairwise import PAIRWISE_KERNEL_FUNCTIONS, pairwise_kernels
 from ..utils import _format_bytes, _log_array, check_array
 from .k_means import KMeans
@@ -142,7 +143,7 @@ class SpectralClustering(BaseEstimator, ClusterMixin):
         eigen_solver=None,
         random_state=None,
         n_init=10,
-        gamma=1.,
+        gamma=1.0,
         affinity="rbf",
         n_neighbors=10,
         eigen_tol=0.0,
@@ -189,7 +190,8 @@ class SpectralClustering(BaseEstimator, ClusterMixin):
         if isinstance(self.assign_labels, six.string_types):
             if self.assign_labels == "kmeans":
                 km = KMeans(
-                    n_clusters=n_clusters, random_state=rng.randint(2 ** 32 - 1)
+                    n_clusters=n_clusters,
+                    random_state=draw_seed(rng, 2 ** 32 - 1, dtype="uint"),
                 )
             elif self.assign_labels == "sklearn-kmeans":
                 km = sklearn.cluster.KMeans(n_clusters=n_clusters, random_state=rng)