+++ dask_ml/preprocessing/__init__.py
@@ -1,9 +1,5 @@
 """Utilties for Preprocessing data.
 """
-from .._compat import SK_VERSION
-from packaging.version import parse
-
-
 from .data import (
     StandardScaler,
     MinMaxScaler,
@@ -12,8 +8,10 @@ from .data import (
     Categorizer,
     DummyEncoder,
     OrdinalEncoder,
+    PolynomialFeatures,
 )
 from .label import LabelEncoder
+from ._encoders import OneHotEncoder
 
 
 __all__ = [
@@ -25,12 +23,6 @@ __all__ = [
     "DummyEncoder",
     "OrdinalEncoder",
     "LabelEncoder",
+    "OneHotEncoder",
+    "PolynomialFeatures",
 ]
-
-if SK_VERSION >= parse("0.20.0.dev0"):
-    from ._encoders import OneHotEncoder  # noqa
-
-    __all__.append("OneHotEncoder")
-
-del SK_VERSION
-del parse