+++ tests/test_parallel_post_fit.py
@@ -1,13 +1,13 @@
 import dask
 import dask.array as da
 import dask.dataframe as dd
+import numpy as np
 import pytest
 import sklearn.datasets
 from sklearn.decomposition import PCA
 from sklearn.ensemble import GradientBoostingClassifier
 from sklearn.linear_model import LinearRegression, LogisticRegression
 
-from dask_ml._compat import SK_GE_020
 from dask_ml.datasets import make_classification
 from dask_ml.utils import assert_eq_ar, assert_estimator_equal
 from dask_ml.wrappers import ParallelPostFit
@@ -36,6 +36,16 @@ def test_no_method_raises():
         clf.predict_proba(X)
 
     assert m.match("The wrapped estimator (.|\n)* 'predict_proba' method.")
+
+
+def test_laziness():
+    clf = ParallelPostFit(LinearRegression())
+    X, y = make_classification(chunks=50)
+    clf.fit(X, y)
+
+    x = clf.score(X, y, compute=False)
+    assert dask.is_dask_collection(x)
+    assert 0 < x.compute() < 1
 
 
 @pytest.mark.parametrize("kind", ["numpy", "dask.dataframe", "dask.array"])
@@ -93,12 +103,8 @@ def test_multiclass():
     X = da.from_array(X, chunks=50)
     y = da.from_array(y, chunks=50)
 
-    if SK_GE_020:
-        kwargs = {"multi_class": "auto"}
-    else:
-        kwargs = {}
     clf = ParallelPostFit(
-        LogisticRegression(random_state=0, n_jobs=1, solver="lbfgs", **kwargs)
+        LogisticRegression(random_state=0, n_jobs=1, solver="lbfgs", multi_class="auto")
     )
 
     clf.fit(X, y)
@@ -113,3 +119,19 @@ def test_multiclass():
 
     assert isinstance(result, da.Array)
     assert_eq_ar(result, expected)
+
+
+def test_auto_rechunk():
+    clf = ParallelPostFit(GradientBoostingClassifier())
+    X, y = make_classification(n_samples=1000, n_features=20, chunks=100)
+    X = X.rechunk({0: 100, 1: 10})
+    clf.fit(X, y)
+
+    assert clf.predict(X).compute().shape == (1000,)
+    assert clf.predict_proba(X).compute().shape == (1000, 2)
+    assert clf.score(X, y) == clf.score(X.compute(), y.compute())
+
+    X, y = make_classification(n_samples=1000, n_features=20, chunks=100)
+    X = X.rechunk({0: 100, 1: 10})
+    X._chunks = (tuple(np.nan for _ in X.chunks[0]), X.chunks[1])
+    clf.predict(X)